Modularized the codebase
