+++ code/classes/pages/walls_design_page.py
@@ -1,23 +1,25 @@
-from PyQt5.QtWidgets import QWizardPage, QVBoxLayout, QHBoxLayout, QPushButton, QLabel, QLineEdit, QListWidget, QMessageBox, QGraphicsView, QGraphicsScene, QGraphicsLineItem
-from PyQt5.QtCore import Qt, QLineF, QPointF, QEvent
-from PyQt5.QtGui import QPen, QColor
+from PyQt5.QtWidgets import QWizardPage, QHBoxLayout, QVBoxLayout, QPushButton, QLineEdit, QListWidget, QMessageBox, QWidget
+from PyQt5.QtCore import Qt, QEvent, QPointF
+from PyQt5.QtGui import QColor
+from classes.zoomable_graphics_view import ZoomableGraphicsView
 import os
 import shutil
 from xml.etree import ElementTree as ET
+from utils.config import WORLDS_GAZEBO_DIR
 
 class WallsDesignPage(QWizardPage):
-    def __init__(self):
+    def __init__(self, scene):
         super().__init__()
         self.setTitle("Design Walls")
 
         self.registerField("world_name", self)
         self.registerField("wall_list", self)
         self.world_manager = None
-        self.wall_items = {}
+        self.scene = scene
 
         layout = QHBoxLayout()
 
-        # Left section (30%)
+        left_widget = QWidget()
         left_layout = QVBoxLayout()
         self.create_world_button = QPushButton("Create New World")
         self.create_world_button.clicked.connect(self.create_new_world)
@@ -53,65 +55,59 @@ class WallsDesignPage(QWizardPage):
         self.apply_button = QPushButton("Apply and Preview")
         self.apply_button.clicked.connect(self.apply_changes)
         left_layout.addWidget(self.apply_button)
+        left_widget.setLayout(left_layout)
 
-        layout.addLayout(left_layout, 30)
-
-        # Right section (70%) - Canvas
-        self.scene = QGraphicsScene()
-        self.view = QGraphicsView(self.scene)
+        self.view = ZoomableGraphicsView(self.scene)
         self.view.setBackgroundBrush(QColor("white"))
         self.view.installEventFilter(self)
-        layout.addWidget(self.view, 70)
+
+        # Set size constraints: canvas at 70% of content width, left panel takes the rest
+        window_width = 1500  # Default wizard width
+        canvas_width = int(window_width * 0.7)  # 70% of window width
+        self.view.setMinimumWidth(int(800 * 0.7))  # Minimum canvas width (70% of 800px window)
+        self.view.setMaximumWidth(canvas_width)  # Initial canvas width
+        left_widget.setMinimumWidth(150)  # Minimum left panel width for usability
+        left_widget.setMaximumWidth(window_width - canvas_width - 260)  # Initial left panel width
+
+        layout.addWidget(left_widget)
+        layout.addWidget(self.view)
 
         self.setLayout(layout)
-
-        # Canvas interaction
-        self.drawing = False
-        self.start_point = None
-        self.current_line = None
 
     def initializePage(self):
         self.world_manager = self.wizard().world_manager
 
+    def snap_to_grid(self, point, grid_spacing=10):
+        x = round(point.x() / grid_spacing) * grid_spacing
+        y = round(point.y() / grid_spacing) * grid_spacing
+        return QPointF(x, y)
+
     def eventFilter(self, obj, event):
-        if obj == self.view:
-            print("Event:", event.type())
-            if event.type() == QEvent.MouseButtonPress and event.button() == Qt.LeftButton and self.world_manager:
-                self.drawing = True
-                self.start_point = self.view.mapToScene(event.pos())
-                self.current_line = QGraphicsLineItem(QLineF(self.start_point, self.start_point))
-                self.current_line.setPen(QPen(Qt.black, 2))
-                self.scene.addItem(self.current_line)
-                return True
-            elif event.type() == QEvent.MouseMove and self.drawing and self.current_line:
-                end_point = self.view.mapToScene(event.pos())
-                self.current_line.setLine(QLineF(self.start_point, end_point))
-                return True
-            elif event.type() == QEvent.MouseButtonRelease and event.button() == Qt.LeftButton and self.drawing and self.world_manager:
-                self.drawing = False
-                end_point = self.view.mapToScene(event.pos())
-                try:
-                    width = float(self.width_input.text() or 0.1)
-                    height = float(self.height_input.text() or 1.0)
-                except ValueError:
-                    width, height = 0.1, 1.0
-                color = self.color_input.text() or "Gray"
-                wall_name = f"wall_{len(self.world_manager.models) + 1}"
-                wall = {
-                    "name": wall_name,
-                    "type": "wall",
-                    "properties": {
-                        "start": (self.start_point.x() / 100, self.start_point.y() / 100),
-                        "end": (end_point.x() / 100, end_point.y() / 100),
-                        "width": width,
-                        "height": height,
-                        "color": color
-                    },
-                    "status": "new"
-                }
-                self.world_manager.add_model(wall)
-                self.wall_list.addItem(wall_name)
-                self.wall_items[wall_name] = self.current_line
+        if obj == self.view and self.world_manager:
+            if event.type() == QEvent.MouseButtonPress and event.button() == Qt.LeftButton:
+                if not hasattr(self, 'start_point'):
+                    clicked_point = self.view.mapToScene(event.pos())
+                    self.start_point = self.snap_to_grid(clicked_point)
+                else:
+                    clicked_point = self.view.mapToScene(event.pos())
+                    end_point = self.snap_to_grid(clicked_point)
+                    wall_name = f"wall_{len(self.world_manager.models) + 1}"
+                    wall = {
+                        "name": wall_name,
+                        "type": "wall",
+                        "properties": {
+                            "start": (self.start_point.x() / 100, -self.start_point.y() / 100),
+                            "end": (end_point.x() / 100, -end_point.y() / 100),
+                            "width": float(self.width_input.text() or 0.1),
+                            "height": float(self.height_input.text() or 1.0),
+                            "color": self.color_input.text() or "Gray"
+                        },
+                        "status": "new"
+                    }
+                    self.world_manager.add_model(wall)
+                    self.wall_list.addItem(wall_name)
+                    self.wizard().refresh_canvas(self.scene)
+                    del self.start_point
                 return True
         return super().eventFilter(obj, event)
 
@@ -124,18 +120,15 @@ class WallsDesignPage(QWizardPage):
             QMessageBox.warning(self, "Error", "Please enter a valid world name.")
             return
         try:
-            # Use WorldManager's path for empty_world.sdf
-            empty_world_path = os.path.join(self.world_manager.base_dir, "worlds", "gazebo", self.world_manager.version, "empty_world.sdf")
-            print(f"Copying empty world from: {empty_world_path}")  # Debug
+            empty_world_path = os.path.join(WORLDS_GAZEBO_DIR, self.world_manager.version, "empty_world.sdf")
+            print(f"Copying empty world from: {empty_world_path}")
             if not os.path.exists(empty_world_path):
                 raise FileNotFoundError(f"Empty world file not found: {empty_world_path}")
-            new_world_path = os.path.join(self.world_manager.base_dir, "worlds", "gazebo", self.world_manager.version, f"{world_name}.sdf")
-            print(f"Creating new world at: {new_world_path}")  # Debug
+            new_world_path = os.path.join(WORLDS_GAZEBO_DIR, self.world_manager.version, f"{world_name}.sdf")
+            print(f"Creating new world at: {new_world_path}")
 
-            # Copy the empty_world.sdf to the new world file
             shutil.copyfile(empty_world_path, new_world_path)
 
-            # Update the world name in the SDF file
             tree = ET.parse(new_world_path)
             root = tree.getroot()
             world_elem = root.find("world")
@@ -145,13 +138,12 @@ class WallsDesignPage(QWizardPage):
                 raise ValueError("SDF file does not contain a <world> element")
             tree.write(new_world_path, encoding="utf-8", xml_declaration=True)
 
-            # Load the new world using WorldManager
             self.world_manager.load_world(world_name)
 
             self.wall_list.clear()
-            self.scene.clear()
-            self.wall_items = {}
+            self.wizard().refresh_canvas(self.scene)
             QMessageBox.information(self, "Success", f"Created and loaded new world: {world_name}")
+            self.completeChanged.emit()
         except Exception as e:
             QMessageBox.critical(self, "Error", f"Failed to create world: {str(e)}")
 
@@ -166,17 +158,12 @@ class WallsDesignPage(QWizardPage):
         try:
             self.world_manager.load_world(world_name)
             self.wall_list.clear()
-            self.scene.clear()
-            self.wall_items = {}
+            self.wizard().refresh_canvas(self.scene)
             for model in self.world_manager.models:
                 if model["type"] == "wall":
                     self.wall_list.addItem(model["name"])
-                    start = QPointF(model["properties"]["start"][0] * 100, model["properties"]["start"][1] * 100)
-                    end = QPointF(model["properties"]["end"][0] * 100, model["properties"]["end"][1] * 100)
-                    line = QGraphicsLineItem(QLineF(start, end))
-                    self.scene.addItem(line)
-                    self.wall_items[model["name"]] = line
             QMessageBox.information(self, "Success", f"Loaded world: {world_name}")
+            self.completeChanged.emit()
         except FileNotFoundError as e:
             QMessageBox.critical(self, "Error", str(e))
         except Exception as e:
@@ -189,9 +176,11 @@ class WallsDesignPage(QWizardPage):
         selected = self.wall_list.currentItem()
         if selected:
             wall_name = selected.text()
-            if wall_name in self.wall_items:
-                self.scene.removeItem(self.wall_items[wall_name])
-                del self.wall_items[wall_name]
+            if wall_name in self.wizard().wall_items:
+                line, text = self.wizard().wall_items[wall_name]
+                self.scene.removeItem(line)
+                self.scene.removeItem(text)
+                del self.wizard().wall_items[wall_name]
             for model in self.world_manager.models:
                 if model["name"] == wall_name:
                     model["status"] = "removed"
@@ -204,6 +193,7 @@ class WallsDesignPage(QWizardPage):
             return
         try:
             self.world_manager.apply_changes()
+            self.wizard().refresh_canvas(self.scene)
             QMessageBox.information(self, "Success", "Changes applied successfully.")
         except Exception as e:
             QMessageBox.critical(self, "Error", f"Failed to apply changes: {str(e)}")
